# docs/api/turn-2-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/businesses/{businessId}/employees/{employeeId}/services": {
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Set which services this employee performs",
                "operationId": "setServices",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetServices"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "List the businesses the caller belongs to",
                "operationId": "listMine",
                "responses": {
                    "200": {
                        "description": "The caller's businesses, possibly empty",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BusinessResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Create a business; the caller becomes its owner",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBusinessRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed or unknown timezone",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/businesses/{businessId}/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "List employees",
                "operationId": "list_1",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EmployeeResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Add an employee",
                "operationId": "create_2",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEmployee"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not a member of this business",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/businesses/{businessId}/employees/{employeeId}/working-hours": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "List an employee's working windows",
                "operationId": "listWorkingHours",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkingHoursResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Add a working window; two on one weekday express a break",
                "operationId": "addWorkingHours",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateWorkingHours"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkingHoursResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "The window does not end after it starts",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkingHoursResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not a member of this business",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkingHoursResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such employee here",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkingHoursResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/businesses/{businessId}": {
            "get": {
                "tags": [
                    "Businesses"
                ],
                "summary": "Read one business the caller belongs to",
                "operationId": "get",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The business",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not a member \u2014 returned identically when the business does not exist",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/businesses/{businessId}/working-hours/{workingHoursId}": {
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Remove a working window",
                "operationId": "deleteWorkingHours",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "workingHoursId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed"
                    },
                    "403": {
                        "description": "Not a member of this business"
                    },
                    "404": {
                        "description": "No such window here"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/businesses/{businessId}/employees/{employeeId}": {
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Remove an employee",
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "businessId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "employeeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed"
                    },
                    "403": {
                        "description": "Not a member of this business"
                    },
                    "404": {
                        "description": "No such employee here"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SetServices": {
                "type": "object",
                "properties": {
                    "serviceIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "uniqueItems": true
                    }
                },
                "required": [
                    "serviceIds"
                ]
            },
            "EmployeeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "serviceIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                }
            },
            "CreateBusinessRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    },
                    "timezone": {
                        "type": "string",
                        "maxLength": 64,
                        "minLength": 0
                    }
                }
            },
            "BusinessResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    }
                }
            },
...
            "CreateEmployee": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "maxLength": 120,
                        "minLength": 0
                    }
                }
            },
            "CreateWorkingHours": {
                "type": "object",
                "properties": {
                    "weekday": {
                        "type": "string",
                        "enum": [
                            "MONDAY",
                            "TUESDAY",
                            "WEDNESDAY",
                            "THURSDAY",
                            "FRIDAY",
                            "SATURDAY",
                            "SUNDAY"
                        ]
                    },
                    "startsAt": {
                        "type": "string",
                        "example": "14:30:00"
                    },
                    "endsAt": {
                        "type": "string",
                        "example": "14:30:00"
                    }
                },
                "required": [
                    "endsAt",
                    "startsAt",
                    "weekday"
                ]
            },
            "WorkingHoursResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "weekday": {
                        "type": "string",
                        "enum": [
                            "MONDAY",
                            "TUESDAY",
                            "WEDNESDAY",
                            "THURSDAY",
                            "FRIDAY",
                            "SATURDAY",
                            "SUNDAY"
                        ]
                    },
                    "startsAt": {
                        "type": "string",
                        "example": "14:30:00"
                    },
                    "endsAt": {
                        "type": "string",
                        "example": "14:30:00"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}